Cleanup typescript
