+++ src/modules/household/components/householdActions/SelectClientsStep.tsx
@@ -3,8 +3,8 @@ import { Paper, Stack } from '@mui/material';
 import { ReactNode, useCallback, useMemo } from 'react';
 import GenericTable from '@/components/elements/table/GenericTable';
 import { ColumnDef } from '@/components/elements/table/types';
+import { WITH_ENROLLMENT_COLUMNS } from '@/modules/enrollment/columns/enrollmentColumns';
 import { clientBriefName, sortHouseholdMembers } from '@/modules/hmis/hmisUtil';
-import { WITH_ENROLLMENT_COLUMNS } from '@/modules/projects/components/tables/ProjectClientEnrollmentsTable';
 import { ENROLLMENT_RELATIONSHIP_COL } from '@/modules/projects/components/tables/ProjectHouseholdsTable';
 import { CLIENT_COLUMNS } from '@/modules/search/components/ClientSearch';
 import { HouseholdClientFieldsFragment } from '@/types/gqlTypes';
